Add files via upload
Heuristic search algorithm for optimizing a communications system. We proposed to maximize the length of the fiber optics and minimize power, keeping the beams' shape as homogeneous as possible. The hyperlatin cube sampling technique for the mesh generation was implemented. These were passed into the non linear Schrödringer's equation and was solved using spectral methods. To determine the best value, two criteria functions were implementes, the first one counted how many points were above a given tolerance, so the system could recognize if it was reading a one or zero, the second one was an invented economic cost function. It prints at the end the best found values.