Add files via upload
The project consisted of predictive maintainace, I was given a dataset, and in the code I preprocess the data, and then I applied non supervised techniques for generating labeld data and finally apply supervised learning for developing the model. Due to the attributes of the data, I was only able to determine who was using the motor.